// src/app/model/product.model.ts
export class Product{
    id: number = 0;
    name: string = '';
    email: string = '';
    phone: number = 0;
    address: string = '';
    productImage:string = '';
    productName: string = '';
    productBrand: string = '';
    productCategory: string = '';
    productPrice: number = 0;
    productDescription: string = '';
    date: string = '';
    showEditOption: boolean = false;
    
}